circulation: support non-circulating libraries
The transaction locations of a library is selected in the list of pickup
locations of the library.  This causes issues with libraries that have
no pickup location setted.

With this commit, for libraries with no pickup locations the system
considers the first library location as transaction location for
circulation transactions. These libraries can now checkout items
directly from other libraries and once checked-in back they go
in_transit.
It is still mandatory to setup library hours and notification
settings before checkout items.

* Upgrades to invenio-circulation v1.0.0a34.
* Supports checkout of items of libraries with no opening hours externally.
* Fixes problem when in transit flash message was always displaying the pickup library.
* Fixes situation A of the issue 2419.
* Supports checkout of items of libraries with no pickup locations.
* Closes #2367.
* Adds a non-circulating library to units testing.

Co-Authored-by: Aly Badr <aly.badr@rero.ch>